[LOGIN, LOGOUT] Fixed user session when logging in and added a logout feature to delete user session
